Make float to double conversions for explicit string conversions also explicit and this better visible. Fixes compilation issues with DMC
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@4199 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CAttributeImpl.h
@@ -168,7 +168,7 @@ public:
 
 	virtual core::stringw getStringW()
 	{
-		return core::stringw(Value);
+		return core::stringw((double)Value);
 	}
 
 	virtual void setInt(s32 intValue)
@@ -1734,11 +1734,11 @@ public:
 	{
 		if (IsStringW)
 		{
-			ValueW = core::stringw(floatValue);
+			ValueW = core::stringw((double)floatValue);
 		}
 		else
 		{
-			Value = core::stringc(floatValue);
+			Value = core::stringc((double)floatValue);
 		}
 	};
